main.py
#!/usr/bin/env python3
"""
Airship Zero - Realistic Airship Flight Simulator
Entry point for the application.
"""

import sys
import os

def main():
    """Main entry point for Airship Zero."""
    print("🎈 Airship Zero - Agentic Coding Challenge")
    print("=" * 45)
    print()
    print("Welcome to the Airship Zero development environment!")
    print("This is the starting point for your AI-powered coding challenge.")
    print()
    print("📋 Challenge Status: READY TO START")
    print("- Specifications: ✓ Complete (see data-model.md)")
    print("- Dependencies: ✓ Configured")  
    print("- Project structure: ✓ Ready")
    print("- AI agent target: 🎯 Full flight simulator")
    print()
    print("🤖 Next steps for your AI agents:")
    print("1. Read data-model.md (3,500+ lines of detailed specs)")
    print("2. Implement core game loop with pygame")
    print("3. Create basic UI framework") 
    print("4. Add navigation system")
    print("5. Develop engine simulation")
    print("6. Build all the complex systems!")
    print()
    print("💡 This is designed to be AI-agent friendly:")
    print("- Function-based architecture (no complex inheritance)")
    print("- Single JSON game state (easy to understand)")
    print("- Clear, detailed specifications")
    print("- Modular, independent systems")
    print()
    print()
    print("🔧 Dependency Check:")
    
    # Basic dependency check
    try:
        import pygame
        print("✓ Pygame available - Ready for game development!")
    except ImportError:
        print("✗ Pygame not found - Install dependencies first")
        
    try:
        import numpy
        print("✓ NumPy available - Math calculations ready!")
    except ImportError:
        print("✗ NumPy not found - Install dependencies first")
        
    try:
        import PIL
        print("✓ Pillow available - Image processing ready!")
    except ImportError:
        print("✗ Pillow not found - Install dependencies first")
        
    try:
        import markdown
        print("✓ Markdown available - Book system ready!")
    except ImportError:
        print("✗ Markdown not found - Install dependencies first")
    
    print()
    print("🚀 Ready for your AI agents to build an amazing flight simulator!")
    print("Good luck with the challenge!")
    return 0

if __name__ == "__main__":
    sys.exit(main())
