pyproject.toml
[project]
name = "airshipzero"
version = "0.1.0"
description = "Realistic airship flight simulator with complex systems management"
authors = [{name = "Timeless Prototype"}]
requires-python = ">=3.12"
dependencies = [
    "pygame>=2.5.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "markdown>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
airshipzero = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
exclude = [
    "/.git",
    "/.gitignore", 
    "/.vscode",
    "/tests",
    "**/__pycache__",
    "*.pyc"
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0", 
    "ruff>=0.1.0",
]
